keys in ReviewForm fixed; header and card component renamed
